Added a home page button and more CSS
Added a button to tuffytalks.html that links to the Login.html page. Added functionality to the Login.html button to link back to tuffytalks.html. Added more CSS to make the buttons look better. The Login.html button still does not store the username and password anywhere, or log the user into their account.